requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
websockets==12.0
aiofiles==23.2.1
python-dotenv==1.0.0
gunicorn==21.2.0
toml==0.10.2
tqdm